tests relocated
