+++ draco-oxide/core/src/codec/attribute/sequence.rs
@@ -126,186 +126,3 @@ where
         self.out
     }
 }
-
-/* crate-split: round-trip test needs encoder+io; relocate to encoder crate
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::types::ConfigType;
-    use crate::encode::connectivity::ConnectivityEncoderOutput;
-    use crate::{encode::connectivity::encode_connectivity, io::obj::load_obj};
-
-    #[test]
-    fn test_traverser() {
-        let mut mesh = load_obj("../tests/data/tetrahedron.obj").unwrap();
-        let out: crate::encode::connectivity::ConnectivityEncoderOutput<'_> = encode_connectivity(
-            &mesh.faces,
-            &mut mesh.attributes,
-            &mut Vec::new(),
-            &crate::encode::Config::default(),
-        )
-        .unwrap();
-
-        let (ct, corners) = if let ConnectivityEncoderOutput::Edgebreaker(edgebreaker_out) = out {
-            (
-                edgebreaker_out.corner_table,
-                edgebreaker_out.corners_of_edgebreaker,
-            )
-        } else {
-            panic!("Expected Edgebreaker Output");
-        };
-
-        let ct_pos = ct.universal_corner_table();
-        let sequence_points = Traverser::new(ct_pos, corners.clone())
-            .compute_seqeunce()
-            .iter()
-            .map(|c| ct_pos.point_idx(*c))
-            .collect::<Vec<_>>();
-        assert_eq!(
-            sequence_points
-                .into_iter()
-                .map(|c| usize::from(c))
-                .collect::<Vec<_>>(),
-            vec![3, 1, 0, 2]
-        );
-
-        let ct_nor = &ct.attribute_corner_table(1).unwrap();
-        let sequence_normals = Traverser::new(ct_nor, corners.clone())
-            .compute_seqeunce()
-            .iter()
-            .map(|c| ct_nor.point_idx(*c))
-            .collect::<Vec<_>>();
-        assert_eq!(
-            sequence_normals
-                .into_iter()
-                .map(|c| usize::from(c))
-                .collect::<Vec<_>>(),
-            vec![3, 1, 0, 2]
-        );
-
-        let ct_tex = &ct.attribute_corner_table(2).unwrap();
-        let sequence_tex_coords = Traverser::new(ct_tex, corners)
-            .compute_seqeunce()
-            .iter()
-            .map(|c| ct_tex.point_idx(*c))
-            .collect::<Vec<_>>();
-        assert_eq!(
-            sequence_tex_coords
-                .into_iter()
-                .map(|c| usize::from(c))
-                .collect::<Vec<_>>(),
-            vec![3, 1, 0, 2, 5, 4]
-        );
-    }
-
-    /// FNV-1a over the little-endian bytes of the point-index sequence.
-    /// Deterministic and toolchain-independent, unlike DefaultHasher.
-    fn digest(seq: &[usize]) -> u64 {
-        let mut h: u64 = 0xcbf2_9ce4_8422_2325;
-        for &v in seq {
-            for b in (v as u64).to_le_bytes() {
-                h ^= b as u64;
-                h = h.wrapping_mul(0x0000_0100_0000_01b3);
-            }
-        }
-        h
-    }
-
-    /// Computes (attr_idx, sequence_len, digest) for the universal corner table
-    /// (attr 0) and every attribute corner table of `mesh`. The digest captures
-    /// the exact `Vec<CornerIdx>` traversal order via point indices — this is the
-    /// shared encoder/decoder symmetry that must stay byte-identical.
-    fn sequence_fingerprints(path: &str) -> Vec<(usize, usize, u64)> {
-        let mut mesh = load_obj(path).unwrap();
-        let out = encode_connectivity(
-            &mesh.faces,
-            &mut mesh.attributes,
-            &mut Vec::new(),
-            &crate::encode::Config::default(),
-        )
-        .unwrap();
-
-        let (ct, corners) = if let ConnectivityEncoderOutput::Edgebreaker(eb) = out {
-            (eb.corner_table, eb.corners_of_edgebreaker)
-        } else {
-            panic!("Expected Edgebreaker Output for {path}");
-        };
-
-        let mut fps = Vec::new();
-
-        let ct_pos = ct.universal_corner_table();
-        let seq: Vec<usize> = Traverser::new(ct_pos, corners.clone())
-            .compute_seqeunce()
-            .iter()
-            .map(|c| usize::from(ct_pos.point_idx(*c)))
-            .collect();
-        fps.push((0, seq.len(), digest(&seq)));
-
-        let mut attr_idx = 1;
-        while let Some(ct_attr) = ct.attribute_corner_table(attr_idx) {
-            let seq: Vec<usize> = Traverser::new(&ct_attr, corners.clone())
-                .compute_seqeunce()
-                .iter()
-                .map(|c| usize::from(ct_attr.point_idx(*c)))
-                .collect();
-            fps.push((attr_idx, seq.len(), digest(&seq)));
-            attr_idx += 1;
-        }
-
-        fps
-    }
-
-    /// Byte-identical oracle for `compute_sequence`. The expected fingerprints
-    /// were captured from the pre-optimization implementation; any change that
-    /// alters the traversal order on these meshes (boundaries, handles) trips
-    /// this test. torus.obj carries topological handles, which is exactly the
-    /// case the handle-detection scan-and-remove blocks exist to handle.
-    #[test]
-    fn oracle_compute_sequence() {
-        let cases: &[(&str, &[(usize, usize, u64)])] = &[
-            ("../tests/data/tetrahedron.obj", EXPECT_TETRAHEDRON),
-            ("../tests/data/sphere.obj", EXPECT_SPHERE),
-            (
-                "../tests/data/punctured_sphere.obj",
-                EXPECT_PUNCTURED_SPHERE,
-            ),
-            ("../tests/data/torus.obj", EXPECT_TORUS),
-            ("../tests/data/bunny.obj", EXPECT_BUNNY),
-        ];
-
-        let dump = std::env::var("DUMP_FINGERPRINTS").is_ok();
-        for (path, expected) in cases {
-            let got = sequence_fingerprints(path);
-            if dump {
-                eprintln!("{path} => {got:?}");
-                continue;
-            }
-            assert_eq!(
-                &got[..],
-                *expected,
-                "compute_sequence output changed for {path}"
-            );
-        }
-    }
-
-    // Captured from the pre-optimization implementation. Format: (attr_idx, len, fnv1a_digest).
-    const EXPECT_TETRAHEDRON: &[(usize, usize, u64)] = &[
-        (0, 4, 18054049684469353541),
-        (1, 4, 18054049684469353541),
-        (2, 6, 3159456026337658052),
-    ];
-    const EXPECT_SPHERE: &[(usize, usize, u64)] = &[
-        (0, 114, 17737425019064467876),
-        (1, 114, 17737425019064467876),
-    ];
-    const EXPECT_PUNCTURED_SPHERE: &[(usize, usize, u64)] = &[
-        (0, 114, 17132826066695074116),
-        (1, 114, 17132826066695074116),
-    ];
-    const EXPECT_TORUS: &[(usize, usize, u64)] = &[(0, 2051, 930682351741064974)];
-    const EXPECT_BUNNY: &[(usize, usize, u64)] = &[
-        (0, 34834, 3080192193140594432),
-        (1, 34834, 3080192193140594432),
-    ];
-}
-*/
