+++ draco-oxide/core/src/corner_table/attribute_corner_table.rs
@@ -5,12 +5,12 @@ use crate::types::{AttributeValueIdx, CornerIdx, VecVertexIdx, VertexIdx};
 
 #[derive(Debug, Clone)]
 pub struct AttributeCornerTable {
-    corner_to_vertex: Vec<VertexIdx>,
-    vertex_to_attribute_map: VecVertexIdx<AttributeValueIdx>,
+    pub corner_to_vertex: Vec<VertexIdx>,
+    pub vertex_to_attribute_map: VecVertexIdx<AttributeValueIdx>,
     /// A vector that indicates whether the edge opposite to the corner is a seam edge.
-    is_edge_on_seam: Vec<bool>,
-    is_vertex_on_seam: Vec<bool>,
-    left_most_corners: Vec<CornerIdx>,
+    pub is_edge_on_seam: Vec<bool>,
+    pub is_vertex_on_seam: Vec<bool>,
+    pub left_most_corners: Vec<CornerIdx>,
     num_vertices: usize,
 }
 
@@ -191,154 +191,3 @@ impl AttributeCornerTable {
         self.left_most_corners[usize::from(vertex)]
     }
 }
-
-/* crate-split: integration test needs io (load_obj); relocate to encoder crate
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::{io::obj::load_obj, prelude::AttributeType};
-    #[test]
-    fn test_no_att_seam() {
-        // read the test data from a corner table
-
-        let mut mesh = load_obj("../tests/data/sphere.obj").unwrap();
-        let faces = mesh.faces;
-
-        let att = mesh
-            .attributes
-            .iter()
-            .find(|att| att.get_attribute_type() == AttributeType::Position)
-            .unwrap();
-
-        let corner_table = CornerTable::new(&faces, &att);
-        let att = mesh
-            .attributes
-            .iter_mut()
-            .find(|att| att.get_attribute_type() == AttributeType::Normal)
-            .unwrap();
-        let attr_corner_table = AttributeCornerTable::new(&corner_table, att);
-        assert_eq!(
-            attr_corner_table.num_vertices(),
-            corner_table.num_vertices()
-        );
-        assert_eq!(
-            attr_corner_table.corner_to_vertex.len(),
-            corner_table.num_corners()
-        );
-        assert_eq!(
-            attr_corner_table.vertex_to_attribute_map.len(),
-            corner_table.num_vertices()
-        );
-        assert_eq!(
-            attr_corner_table.left_most_corners.len(),
-            corner_table.num_vertices()
-        );
-        assert_eq!(
-            attr_corner_table.is_edge_on_seam.len(),
-            corner_table.num_corners()
-        );
-        assert_eq!(
-            attr_corner_table.is_vertex_on_seam.len(),
-            corner_table.num_vertices()
-        );
-        assert!(attr_corner_table
-            .is_edge_on_seam
-            .iter()
-            .all(|&x| x == false));
-        assert!(attr_corner_table
-            .is_vertex_on_seam
-            .iter()
-            .all(|&x| x == false));
-        assert!(attr_corner_table
-            .left_most_corners
-            .iter()
-            .all(|&x| usize::from(x) < corner_table.num_corners()));
-        assert!(attr_corner_table
-            .corner_to_vertex
-            .iter()
-            .all(|&x| usize::from(x) < corner_table.num_vertices()));
-
-        // check the opposite corners
-        for c in 0..corner_table.num_corners() {
-            let c = CornerIdx::from(c);
-            assert_eq!(
-                attr_corner_table.opposite(c, &corner_table),
-                corner_table.opposite(c)
-            );
-        }
-
-        // check vertices
-        for c in 0..corner_table.num_corners() {
-            let c = CornerIdx::from(c);
-            assert_eq!(
-                attr_corner_table.vertex_idx(c),
-                corner_table.vertex_idx(c),
-                "attr corner_to_vertex: {:?}\noriginal corner_to_vertex: {:?}",
-                attr_corner_table.corner_to_vertex,
-                corner_table.corner_to_vertex
-            );
-        }
-
-        // no attribute seams, so all edges and vertices are not on a seam.
-        attr_corner_table.is_edge_on_seam.iter().all(|&x| !x);
-        attr_corner_table.is_vertex_on_seam.iter().all(|&x| !x);
-    }
-
-    #[test]
-    fn test_att_seam() {
-        let mut tetrahedron = load_obj("../tests/data/tetrahedron.obj").unwrap();
-        let faces = tetrahedron.faces;
-        let corner_table = CornerTable::new(&faces, &tetrahedron.attributes[0]);
-
-        let tex_att = tetrahedron
-            .attributes
-            .iter_mut()
-            .find(|att| att.get_attribute_type() == AttributeType::TextureCoordinate)
-            .unwrap();
-        let attr_corner_table = AttributeCornerTable::new(&corner_table, tex_att);
-        assert_eq!(
-            attr_corner_table.num_vertices(),
-            corner_table.num_vertices() + 2
-        );
-        assert_eq!(
-            attr_corner_table.corner_to_vertex.len(),
-            corner_table.num_corners()
-        );
-        assert_eq!(attr_corner_table.corner_to_vertex[0], 0.into());
-        assert_eq!(attr_corner_table.swing_left(4.into(), &corner_table), None);
-        assert_eq!(attr_corner_table.swing_right(4.into(), &corner_table), None);
-        assert_eq!(attr_corner_table.swing_left(8.into(), &corner_table), None);
-        assert_eq!(attr_corner_table.swing_right(8.into(), &corner_table), None);
-        assert_eq!(attr_corner_table.swing_left(10.into(), &corner_table), None);
-        assert_eq!(
-            attr_corner_table.swing_right(10.into(), &corner_table),
-            None
-        );
-        let seam_edge_corners = [3, 5, 6, 7, 9, 11];
-        for c in seam_edge_corners {
-            let c = CornerIdx::from(c);
-            assert!(
-                attr_corner_table.is_corner_opposite_to_seam_edge(c),
-                "Corner {:?} is not opposite to a seam edge, but it should be. is_edge_on_seam: {:?}",
-                c, attr_corner_table.is_edge_on_seam
-            )
-        }
-        let left_most_corners = [6, 5, 11, 10, 8, 4];
-        for (v, left_most_corner) in left_most_corners.into_iter().enumerate() {
-            let v = VertexIdx::from(v);
-            let left_most_corner = CornerIdx::from(left_most_corner);
-            assert_eq!(
-                attr_corner_table.left_most_corner(v), left_most_corner,
-                "Left most corner for vertex {:?} is {:?}, but it should be {:?}. left_most_corners: {:?}",
-                v,
-                attr_corner_table.left_most_corner(v),
-                left_most_corner,
-                attr_corner_table.left_most_corners
-            );
-            assert!(attr_corner_table
-                .swing_left(left_most_corner, &corner_table)
-                .is_none(),);
-        }
-    }
-}
-*/
