+++ draco-oxide/decoder/src/decode/entropy/symbol_coding.rs
@@ -124,99 +124,3 @@ where
     }
     Ok(out)
 }
-
-/* crate-split: round-trip test depends on encode; disabled while decode is WIP
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::encode::entropy::*;
-
-    #[test]
-    fn test_encode_decode_symbols() -> Result<(), Err> {
-        let len = 100;
-        let symbols = (0..len).map(|x| (x*x*x)%23).collect::<Vec<_>>();
-        let mut buffer = Vec::new();
-        symbol_coding::encode_symbols(
-            symbols.clone(),
-            1,
-            SymbolEncodingMethod::LengthCoded,
-            &mut buffer
-        ).unwrap();
-        let mut reader = buffer.into_iter();
-        let decoded_symbols = decode_symbols(
-            len as usize,
-            1,
-            &mut reader
-        )?;
-        assert_eq!(reader.next(), None, "Reader should be empty after decoding all symbols");
-        assert_eq!(decoded_symbols, symbols);
-        Ok(())
-    }
-
-        #[test]
-    fn test_encode_decode_symbols_multi_components() -> Result<(), Err> {
-        let len = 300;
-        let symbols = (0..len).map(|x| (x*x*x)%23).collect::<Vec<_>>();
-        let mut buffer = Vec::new();
-        symbol_coding::encode_symbols(
-            symbols.clone(),
-            3,
-            SymbolEncodingMethod::LengthCoded,
-            &mut buffer
-        ).unwrap();
-        let mut reader = buffer.into_iter();
-        let decoded_symbols = decode_symbols(
-            len as usize,
-            3,
-            &mut reader
-        )?;
-        assert_eq!(reader.next(), None, "Reader should be empty after decoding all symbols");
-        assert_eq!(decoded_symbols, symbols);
-        Ok(())
-    }
-
-    #[test]
-    fn test_encode_decode_symbols_direct_coded() -> Result<(), Err> {
-        let len = 100;
-        let symbols = (0..len).map(|x| (x*x*x)%23).collect::<Vec<_>>();
-        let mut buffer = Vec::new();
-        symbol_coding::encode_symbols(
-            symbols.clone(),
-            1,
-            SymbolEncodingMethod::DirectCoded,
-            &mut buffer
-        ).unwrap();
-        let mut reader = buffer.into_iter();
-        let decoded_symbols = decode_symbols(
-            len as usize,
-            1,
-            &mut reader
-        )?;
-        assert_eq!(reader.next(), None, "Reader should be empty after decoding all symbols");
-        assert_eq!(decoded_symbols, symbols);
-        Ok(())
-    }
-
-    #[test]
-    fn test_encode_decode_symbols_direct_coded_multi_components() -> Result<(), Err> {
-        let len = 300;
-        let symbols = (0..len).map(|x| (x*x*x)%23).collect::<Vec<_>>();
-        let mut buffer = Vec::new();
-        symbol_coding::encode_symbols(
-            symbols.clone(),
-            3,
-            SymbolEncodingMethod::DirectCoded,
-            &mut buffer
-        ).unwrap();
-        let mut reader = buffer.into_iter();
-        let decoded_symbols = decode_symbols(
-            len as usize,
-            3,
-            &mut reader
-        )?;
-        assert_eq!(reader.next(), None, "Reader should be empty after decoding all symbols");
-        assert_eq!(decoded_symbols, symbols);
-        Ok(())
-    }
-}
-*/
